cc2538: ccm: Make it possible to use the interrupt
Using the AES interrupt allows the user process not to waste time polling for
the completion of the operation. This time can be used by the user process to do
something else, or to let the system enter PM0.

Since the system is now free to perform various operations during a crypto
operation, a protection of the crypto resource is added, and PM1+ is prohibited
in order not to stall crypto operations.

Signed-off-by: Benoît Thébaudeau <benoit.thebaudeau@advansee.com>