cc2538: Move SoC data to a dedicated section to save space
Some SoC data requires huge alignments. E.g., the µDMA channel control table has
to be 1024-byte aligned, which can waste up to 1023 bytes of RAM because of the
alignment gap. Even more space can be wasted because this forces the alignment
of the .bss section to be at least 1024 bytes. Hence, such SoC data should be
carefully placed instead of relying on the linker.

One solution could have been to move the .bss section before the .data section,
but such data could also appear in the .data section. Consequently, the best
solution seems to be to create a dedicated section to place such SoC data
requiring huge alignments, and to put this section at the beginning of the SRAM.

In the case of the µDMA channel control table, the data is still zeroed by
udma_init() (instead of also being zeroed as part of .bss).

Signed-off-by: Benoît Thébaudeau <benoit.thebaudeau@advansee.com>